Rename utils to contstants as that is all it contains now
