+++ grpc4bmi/constants.py
@@ -1,2 +1,2 @@
-# grpc max message size is 4Mb
 GRPC_MAX_MESSAGE_LENGTH = 4 * 1024 * 1024
+"""grpc max message size is 4Mb"""
